Move types into their own file
As a result of this change I had to add `typescript/no-unused-vars` rule, so I went ahead and added a bunch more.
