+++ src/.eslintrc.json
@@ -9,7 +9,18 @@
     "plugins": ["typescript"],
     "rules": {
         "no-undef": "off",
-        "camelcase": "off"
+        "camelcase": "off",
+        "typescript/adjacent-overload-signatures": "error",
+        "typescript/class-name-casing": "error",
+        "typescript/interface-name-prefix": "error",
+        "typescript/member-delimiter-style": "error",
+        "typescript/no-unused-vars": "error",
+        "typescript/member-ordering": "error",
+        "typescript/no-angle-bracket-type-assertion": "error",
+        "typescript/no-array-constructor": "error",
+        "typescript/no-empty-interface": "error",
+        "typescript/no-use-before-define": "error",
+        "typescript/type-annotation-spacing": "error"
     },
     "settings": {
         "import/resolver": {
